requirements.txt
lupa
mido
python-rtmidi
sounddevice
numpy
pillow